+++ src/wakepy/core/__init__.py
@@ -21,6 +21,9 @@ from .mode import ActivationWarning as ActivationWarning
 from .mode import ContextAlreadyEnteredError as ContextAlreadyEnteredError
 from .mode import Mode as Mode
 from .mode import ModeExit as ModeExit
+from .mode import current_mode as current_mode
+from .mode import global_modes as global_modes
+from .mode import modecount as modecount
 from .platform import CURRENT_PLATFORM as CURRENT_PLATFORM
 from .registry import get_method as get_method
 from .registry import get_methods as get_methods